Add files via upload
CloudNexus, a dynamic and progressive company, specializes in cutting-edge cloud solutions for global customers. With a team of 600 talented experts, CloudNexus focuses on creating and executing customized cloud-based solutions to meet the evolving needs of businesses across various sectors. As the company prepares to relocate to a larger facility, a new network service must be developed and implemented to ensure strong security and seamless operations in the digital world